Refactor_Flowstepコンポーネントのリファクタリング
