+++ src/resources/js/Components/Flowstep.jsx
@@ -1,18 +1,21 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useDrag } from 'react-dnd';
 import '../../css/Flowstep.css';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchFlowsteps, deleteFlowstepAsync, updateFlowstepName } from '../store/flowstepsSlice'; // Import the async actions
-import { openUpdateFlowstepModal } from '../store/modalSlice';
-import { setSelectedFlowstep, setSelectedMember, setSelectedStepNumber } from '../store/selectedSlice';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrash, faEdit, faSave, faCancel, faPencil } from '@fortawesome/free-solid-svg-icons';
+import { useFlowstep } from '../Hooks/useFlowstep';
 
-const FlowStep = ({ member, flowstep, flowNumber, workflowId }) => {
-    if (!flowstep) {
-        return <div>フローステップのデータがありません</div>;
-    }
+const FlowStep = ({ flowstep }) => {
+    const { 
+        // Local State
+        isDeleting, isEditing, newName, setNewName, isHovered, setIsHovered,
+        // Global State
+          // none 
+        // Event Handler
+        handleDelete, handleEditClick, handleCancel, handleSubmit, handleOpenUpdateFlowstepModal
+    } = useFlowstep(flowstep);
 
+      
     const [{ isDragging }, drag] = useDrag(() => ({
         type: 'FLOWSTEP',
         item: { id: flowstep.id, name: flowstep.name },
@@ -20,56 +23,7 @@ const FlowStep = ({ member, flowstep, flowNumber, workflowId }) => {
             isDragging: monitor.isDragging(),
         }),
     }));
-
-    const [isDeleting, setIsDeleting] = useState(false);
-    const [isEditing, setIsEditing] = useState(false);
-    const [newName, setNewName] = useState(flowstep.name); // Initial value from flowstep.name
-    const [isHovered, setIsHovered] = useState(false); // Hover state
-    const dispatch = useDispatch();
-
-    // Reduxから選択されたメンバーとフローステップの状態を取得
-    const selectedMember = useSelector((state) => state.selected.selectedMember);
-    const selectedStepNumber = useSelector((state) => state.selected.selectedStepNumber);
-    const selectedFlowstep = useSelector((state) => state.selected.selectedFlowstep);
-
-    // デバッグ用ログ
-    console.log('selectedMember on Flowstep.jsx:', selectedMember);
-    console.log('selectedFlowstep on Flowstep.jsx:', selectedFlowstep);
-    console.log('selectedStepNumber on Flowstep.jsx:', selectedStepNumber);
     
-    
-
-    const handleDelete = async () => {
-        setIsDeleting(true); // Set deleting state
-        await dispatch(deleteFlowstepAsync(flowstep.id)); // Dispatch delete action
-        setIsDeleting(false); // Reset deleting state
-    };
-
-    const handleEditClick = () => {
-        setIsEditing(true);
-        setNewName(flowstep.name); // Reset the name to the original value
-    };
-
-    const handleCancel = () => {
-        setIsEditing(false);
-        setNewName(flowstep.name); // Reset the name to the original value
-    };
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        await dispatch(updateFlowstepName({ id: flowstep.id, updatedFlowstep: { name: newName } }));
-        setIsEditing(false); // Close the form after submitting
-        dispatch(fetchFlowsteps(workflowId));
-    };
-
-    const handleOpenUpdateFlowstepModal = (member, flowstep, stepNumber) => {
-        dispatch(setSelectedMember(selectedMember));
-        dispatch(setSelectedStepNumber(stepNumber));
-        dispatch(setSelectedFlowstep(flowstep));
-        dispatch(openUpdateFlowstepModal());
-    };
-
-
     return (
         <div 
             ref={drag}
